# perf(http/file_server): read fileinfo in parallel