part of #3361

Previously, fileInfo was read serially when generating directory listing pages. Change this to read in parallel.
After this PR, the display of the directory listing (page below) is tens of milliseconds faster.

![image](https://user-images.githubusercontent.com/40050810/236527912-9cc11ed5-a593-49e8-8e68-e05a4ce05851.png)

```
// When there are 100 files in the directory:
benchmark      time (avg)             (min … max)       p75       p99      p995
------------------------------------------------- -----------------------------
before     29.6 ms/iter (20.56 ms … 34.89 ms) 31.05 ms 34.89 ms 34.89 ms
after      4.27 ms/iter   (3.51 ms … 6.63 ms)  4.44 ms  5.27 ms  6.63 ms

// When there are 1000 files in the directory:
benchmark      time (avg)             (min … max)       p75       p99      p995
------------------------------------------------- -----------------------------
before     310.22 ms/iter (295.08 ms … 331.81 ms) 317.34 ms 331.81 ms 331.81 ms
after       30.35 ms/iter   (28.24 ms … 32.86 ms)  31.65 ms  32.86 ms  32.86 ms
```

<details>
<summary>benchmark code:</summary>

```ts
// I created 100 files in the testdata/bench directory and run the `deno bench` command.

import { serveDir } from "./file_server.ts";
import { serveDir as serveDirOld } from "./file_server_old.ts";

Deno.bench(async function after() {
  const res = await serveDir(
    new Request("http://localhost:4507/http/testdata/bench/"),
    { quiet: true, showDirListing: true },
  );
  await res.text();
});

Deno.bench(async function before() {
  const res = await serveDirOld(
    new Request("http://localhost:4507/http/testdata/bench/"),
    { quiet: true, showDirListing: true },
  );
  await res.text();
});
```
</details>