@ayame113 2023-05-06
Thank you for your review!  🙇‍♂️

There is a PR on deno_blog that does a similar optimization. If you don't mind, please take a look when you have time.　https://github.com/denoland/deno_blog/pull/116